+++ packages/language-service/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @effect-dynamodb/language-service
 
+## 1.9.5
+
 ## 1.9.4
 
 ## 1.9.3